Fix formatting string assertion failure in the search/replace dialog. Closes #1478.
Originally committed to SVN as r6727.

+++ aegisub/src/dialog_search_replace.cpp
@@ -416,7 +416,7 @@ void SearchReplaceEngine::ReplaceAll() {
 
 	if (count > 0) {
 		context->ass->Commit(_("replace"), AssFile::COMMIT_DIAG_TEXT);
-		wxMessageBox(wxString::Format(_("%i matches were replaced."),count));
+		wxMessageBox(wxString::Format(_("%i matches were replaced."), (int)count));
 	}
 	else {
 		wxMessageBox(_("No matches found."));
